Move agent CSV code to `input/agent.rs`
